+++ src/ui/components/transactions/Transfer/TransferInfo.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { translate, Trans } from 'react-i18next';
 import * as styles from './transfer.styl';
 import { I18N_NAME_SPACE } from '../../../appConfig';
-import { Balance, DateFormat } from '../../ui';
+import { Attachment, Balance, DateFormat } from '../../ui';
 import { getFee, getAmount } from './parseTx';
 import { getMoney } from '../../../utils/converters';
 
@@ -35,9 +35,9 @@ export class TransferInfo extends React.PureComponent<ITransferInfo> {
             
             { tx.attachment ? <div className={`${styles.txRow} ${styles.txRowDescription}`}>
                 <div className="tx-title tag1 basic500">
-                    <Trans i18nKey='transactions.description'>Description</Trans>
+                    <Trans i18nKey='transactions.attachment'>Attachment</Trans>
                 </div>
-                <div className={`${styles.txValue} plate fullwidth`}>{tx.attachment}</div>
+                <Attachment className={`${styles.txValue} plate fullwidth`} attachment={tx.attachment}/>
             </div> : null }
             
             <div className={styles.txRow}>